requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
requests>=2.32.0
mysql-connector-python>=9.0.0
psycopg2-binary>=2.9.9
pytest>=8.3.0
anyio>=4.4.0
